Custom setup for static clusters
