+++ internal/cmd/api/config.go
@@ -2,22 +2,15 @@ package api
 
 import (
 	"context"
-	"fmt"
 	"slices"
-	"strings"
 
+	"github.com/nais/api/internal/kubernetes"
 	"github.com/sethvargo/go-envconfig"
 )
 
-type StaticCluster struct {
-	Name  string
-	Host  string
-	Token string
-}
-
 type k8sConfig struct {
-	Clusters       []string        `env:"KUBERNETES_CLUSTERS"`
-	StaticClusters []StaticCluster `env:"KUBERNETES_CLUSTERS_STATIC"`
+	Clusters       []string                   `env:"KUBERNETES_CLUSTERS"`
+	StaticClusters []kubernetes.StaticCluster `env:"KUBERNETES_CLUSTERS_STATIC"`
 }
 
 func (k *k8sConfig) AllClusterNames() []string {
@@ -80,39 +73,6 @@ func (k *k8sConfig) ClusterList() ClusterList {
 	}
 
 	return clusters
-}
-
-func (c *StaticCluster) EnvDecode(value string) error {
-	if value == "" {
-		return nil
-	}
-
-	parts := strings.Split(value, "|")
-	if len(parts) != 3 {
-		return fmt.Errorf(`invalid static cluster entry: %q. Must be on format "name|host|token"`, value)
-	}
-
-	name := strings.TrimSpace(parts[0])
-	if name == "" {
-		return fmt.Errorf("invalid static cluster entry: %q. Name must not be empty", value)
-	}
-
-	host := strings.TrimSpace(parts[1])
-	if host == "" {
-		return fmt.Errorf("invalid static cluster entry: %q. Host must not be empty", value)
-	}
-
-	token := strings.TrimSpace(parts[2])
-	if token == "" {
-		return fmt.Errorf("invalid static cluster entry: %q. Token must not be empty", value)
-	}
-
-	*c = StaticCluster{
-		Name:  name,
-		Host:  host,
-		Token: token,
-	}
-	return nil
 }
 
 type usersyncConfig struct {
