+++ internal/kubernetes/config.go
@@ -3,6 +3,7 @@ package kubernetes
 import (
 	"fmt"
 	"net/http"
+	"strings"
 
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	"k8s.io/client-go/rest"
@@ -17,7 +18,7 @@ type StaticCluster struct {
 
 type ClusterConfigMap map[string]*rest.Config
 
-func CreateClusterConfigMap(tenant string, clusters []string) (ClusterConfigMap, error) {
+func CreateClusterConfigMap(tenant string, clusters []string, staticClusters []StaticCluster) (ClusterConfigMap, error) {
 	configs := ClusterConfigMap{}
 
 	for _, cluster := range clusters {
@@ -29,13 +30,6 @@ func CreateClusterConfigMap(tenant string, clusters []string) (ClusterConfigMap,
 			WrapTransport: func(rt http.RoundTripper) http.RoundTripper {
 				return otelhttp.NewTransport(rt, otelhttp.WithServerName(cluster))
 			},
-		}
-	}
-
-	staticClusters := make([]StaticCluster, len(clusters))
-	for i, cluster := range clusters {
-		staticClusters[i] = StaticCluster{
-			Name: cluster,
 		}
 	}
 
@@ -62,4 +56,37 @@ func getStaticClusterConfigs(clusters []StaticCluster) ClusterConfigMap {
 		}
 	}
 	return configs
+}
+
+func (c *StaticCluster) EnvDecode(value string) error {
+	if value == "" {
+		return nil
+	}
+
+	parts := strings.Split(value, "|")
+	if len(parts) != 3 {
+		return fmt.Errorf(`invalid static cluster entry: %q. Must be on format "name|host|token"`, value)
+	}
+
+	name := strings.TrimSpace(parts[0])
+	if name == "" {
+		return fmt.Errorf("invalid static cluster entry: %q. Name must not be empty", value)
+	}
+
+	host := strings.TrimSpace(parts[1])
+	if host == "" {
+		return fmt.Errorf("invalid static cluster entry: %q. Host must not be empty", value)
+	}
+
+	token := strings.TrimSpace(parts[2])
+	if token == "" {
+		return fmt.Errorf("invalid static cluster entry: %q. Token must not be empty", value)
+	}
+
+	*c = StaticCluster{
+		Name:  name,
+		Host:  host,
+		Token: token,
+	}
+	return nil
 }
